refactor: migrate provider to jamfplatform-go-sdk v0.2.0
Update all resources, data sources, and actions to work with the
updated SDK. Key changes:

- Remove environment_id provider attribute (SDK consolidated to tenant_id)
- Make tenant_id required for all API operations
- Fix type mismatches from SDK pointer field changes (BlueprintUpdateRequestV1,
  DeviceGroupUpdateRepresentationV1, CBEngineBaselineInfoV1.RuleCount)
- Add missing SDK fields: mdm_capable (device), can_switch_to_enforce
  (benchmark), modified (benchmarks list), reportable/smart_card (rules)
- Remove Sensitive from client_id schema attribute
- Pass tenant_id through to acceptance test helper client
- Regenerate documentation

+++ internal/resources/cbengine/baselines/data_source.go
@@ -135,7 +135,7 @@ func (d *BaselinesDataSource) Read(ctx context.Context, req datasource.ReadReque
 			BaselineID:  types.StringValue(b.BaselineID),
 			Title:       types.StringValue(b.Title),
 			Description: types.StringValue(b.Description),
-			RuleCount:   types.Int64Value(b.RuleCount),
+			RuleCount:   types.Int64Value(int64(b.RuleCount)),
 		})
 	}
